requirements.txt
asgiref==3.9.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.5
django-csp==4.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
whitenoise==6.11.0
